# lib/api-zod/src/generated/types/flight.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alvarez Aviation FOQA Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Flight {
  id: number;
  filename: string;
  aircraftIdent: string;
  product: string;
  softwareVersion: string;
  /** @nullable */
  systemId?: string | null;
  /** @nullable */
  airframeHours?: number | null;
  /** @nullable */
  engineHours?: number | null;
  totalPoints: number;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  /** @nullable */
  maxAltGps?: number | null;
  /** @nullable */
  maxIas?: number | null;
  /** @nullable */
  maxRpm?: number | null;
  /** @nullable */
  maxEgt?: number | null;
  /** @nullable */
  fuelUsed?: number | null;
  uploadedAt: string;
}