# lib/api-zod/src/generated/types/flightPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Alvarez Aviation FOQA Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface FlightPoint {
  id: number;
  flightId: number;
  lclTime: string;
  /** @nullable */
  utcTime?: string | null;
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  /** @nullable */
  altGps?: number | null;
  /** @nullable */
  altP?: number | null;
  /** @nullable */
  ias?: number | null;
  /** @nullable */
  tas?: number | null;
  /** @nullable */
  gndSpd?: number | null;
  /** @nullable */
  trk?: number | null;
  /** @nullable */
  hdg?: number | null;
  /** @nullable */
  pitch?: number | null;
  /** @nullable */
  roll?: number | null;
  /** @nullable */
  vspd?: number | null;
  /** @nullable */
  oat?: number | null;
  /** @nullable */
  e1Rpm?: number | null;
  /** @nullable */
  e1Map?: number | null;
  /** @nullable */
  e1OilT?: number | null;
  /** @nullable */
  e1OilP?: number | null;
  /** @nullable */
  e1Fflow?: number | null;
  /** @nullable */
  volts1?: number | null;
  /** @nullable */
  amps1?: number | null;
  /** @nullable */
  fqty1?: number | null;
  /** @nullable */
  fqty2?: number | null;
  /** @nullable */
  e1Cht1?: number | null;
  /** @nullable */
  e1Cht2?: number | null;
  /** @nullable */
  e1Cht3?: number | null;
  /** @nullable */
  e1Cht4?: number | null;
  /** @nullable */
  e1Cht5?: number | null;
  /** @nullable */
  e1Cht6?: number | null;
  /** @nullable */
  e1Egt1?: number | null;
  /** @nullable */
  e1Egt2?: number | null;
  /** @nullable */
  e1Egt3?: number | null;
  /** @nullable */
  e1Egt4?: number | null;
  /** @nullable */
  e1Egt5?: number | null;
  /** @nullable */
  e1Egt6?: number | null;
  /** @nullable */
  alerts?: string | null;
  /** @nullable */
  baro?: number | null;
}